refactor test folder for StarterPack to introduce subTests and split out SAND and ETH
